Environment_oM: Add Attribute for Plumbing and Fire Protection on Space Object
