+++ Environment_oM/Environment_oM.csproj
@@ -62,21 +62,25 @@
     <Compile Include="Elements\Space.cs" />
     <Compile Include="Fragments\AnalyticalConstruction.cs" />
     <Compile Include="Fragments\LightReflectanceFragment.cs" />
-    <Compile Include="Gains\Acoustics.cs" />
-    <Compile Include="Gains\Enums\EmitterType.cs" />
-    <Compile Include="Gains\Enums\ProfileDay.cs" />
-    <Compile Include="Gains\Enums\ProfileCategory.cs" />
-    <Compile Include="Gains\Enums\ProfileType.cs" />
-    <Compile Include="Gains\Enums\SimulationDayType.cs" />
-    <Compile Include="Gains\Equipment.cs" />
-    <Compile Include="Gains\Exhaust.cs" />
-    <Compile Include="Gains\IGain.cs" />
-    <Compile Include="Gains\Infiltration.cs" />
-    <Compile Include="Gains\Lighting.cs" />
-    <Compile Include="Gains\People.cs" />
-    <Compile Include="Gains\Plug.cs" />
-    <Compile Include="Gains\Pollutant.cs" />
-    <Compile Include="Ventilation\Ventilation.cs" />
+    <Compile Include="SpaceCriteria\Acoustics.cs" />
+    <Compile Include="SpaceCriteria\Enums\EmitterType.cs" />
+    <Compile Include="SpaceCriteria\Enums\ProfileDay.cs" />
+    <Compile Include="SpaceCriteria\Enums\ProfileCategory.cs" />
+    <Compile Include="SpaceCriteria\Enums\HazardType.cs" />
+    <Compile Include="SpaceCriteria\Enums\SprinklerSystemType.cs" />
+    <Compile Include="SpaceCriteria\Enums\ProfileType.cs" />
+    <Compile Include="SpaceCriteria\Enums\SimulationDayType.cs" />
+    <Compile Include="SpaceCriteria\Equipment.cs" />
+    <Compile Include="SpaceCriteria\Exhaust.cs" />
+    <Compile Include="SpaceCriteria\IGain.cs" />
+    <Compile Include="SpaceCriteria\Infiltration.cs" />
+    <Compile Include="SpaceCriteria\Lighting.cs" />
+    <Compile Include="SpaceCriteria\FireProtection.cs" />
+    <Compile Include="SpaceCriteria\Plumbing.cs" />
+    <Compile Include="SpaceCriteria\People.cs" />
+    <Compile Include="SpaceCriteria\Plug.cs" />
+    <Compile Include="SpaceCriteria\Pollutant.cs" />
+    <Compile Include="SpaceCriteria\Ventilation.cs" />
     <Compile Include="MaterialFragments\Enums\Function.cs" />
     <Compile Include="Fragments\BuildingAnalyticalFragment.cs" />
     <Compile Include="Fragments\BuildingContextFragment.cs" />
@@ -91,8 +95,8 @@
     <Compile Include="Fragments\RadiationFragment.cs" />
     <Compile Include="Fragments\SpaceAnalyticalFragment.cs" />
     <Compile Include="Fragments\SpaceContextFragment.cs" />
-    <Compile Include="Gains\Profile.cs" />
-    <Compile Include="Gains\Thermostat.cs" />
+    <Compile Include="SpaceCriteria\Profile.cs" />
+    <Compile Include="SpaceCriteria\Thermostat.cs" />
     <Compile Include="IEnvironmentObject.cs" />
     <Compile Include="MaterialFragments\Enums\Gas.cs" />
     <Compile Include="MaterialFragments\Enums\Roughness.cs" />
