+++ Environment_oM/SpaceCriteria/Acoustics.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
  * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
  *
@@ -20,14 +20,29 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-namespace BH.oM.Environment.Gains
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+using BH.oM.Base;
+using System.ComponentModel;
+
+namespace BH.oM.Environment.SpaceCriteria
+
 {
-    public enum ProfileCategory
+    [Description("The Acoustics object defines the noise criteria for an environments space, whether it's the prescribed ductwork velocity, the noise rating for the space or the dB level.")]
+    public class Acoustics : BHoMObject
     {
-        Undefined,
-        Gain,
-        Humidistat,
-        Thermostat,
+        [Description("The prescribed velocity of the ductwork in the space (with more strict requirements for velocity in bedrooms and theatres (~3 m/s)")]
+        public virtual double DuctVelocity { get; set; } = 0.0;
+
+        [Description("The noise rating (NR), also known as noise criteria (NC) in the US, is the perceived loudness of a space")]
+        public virtual double NoiseRating { get; set; } = 0.0;
+
+        [Description("The Decibel A (dBA) is the sound pressure sensed by the human ear without sensitivity to very low and very high frequencies prescribed for a space")]
+        public virtual double DecibelA { get; set; } = 0.0;
+
     }
 }
-
