+++ Environment_oM/SpaceCriteria/Enums/EmitterType.cs
@@ -20,28 +20,13 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-using BH.oM.Base;
-using System.ComponentModel;
-
-namespace BH.oM.Environment.Gains
+namespace BH.oM.Environment.SpaceCriteria
 {
-    [Description("Plug gains are defined as the amount of sensible heat contributed by electrical receptacle/outlet loads")]
-    public class Plug : BHoMObject, IGain
+    public enum EmitterType
     {
-        [Description("The sensible heat contributed by plugs, which contributes to a rise in temperature with no change in phase.")]
-        public virtual double Sensible { get; set; } = 0.0;
-
-        [Description("Profiles depict the time period (hours per day, days per week) during which plug loads are contributing heat to the space.")]
-        public virtual Profile Profile { get; set; } = new Profile();
-
-        [Description("The radiant fraction depicts the percentage of long wave radiant heat given off by plugs.")]
-        public virtual double RadiantFraction { get; set; } = 0.0;
+        Undefined,
+        Cooling,
+        Heating,
     }
 }
