+++ Environment_oM/SpaceCriteria/Enums/ProfileType.cs
@@ -20,31 +20,14 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-using BH.oM.Base;
-using System.ComponentModel;
-
-namespace BH.oM.Environment.Gains
+namespace BH.oM.Environment.SpaceCriteria
 {
-    [Description("Lighting gains are defined as the amount of heat contributed by light fixtures")]
-    public class Lighting : BHoMObject, IGain
+    public enum ProfileType
     {
-        [Description("The sensible heat contributed by light fixtures, which contributes to a rise in temperature with no change in phase.")]
-        public virtual double Sensible { get; set; } = 0.0;
-
-        [Description("Profiles depict the time period (hours per day, days per week) during which the lighting gain is contributing heat to the space.")]
-        public virtual Profile Profile { get; set; } = new Profile();
-
-        [Description("The radiant fraction depicts the percentage of long wave radiant heat given off by the light fixtures.")]
-        public virtual double RadiantFraction { get; set; } = 0.0;
-
-        [Description("Luminous efficacy is a measure of how well a light fixture produces visible light")]
-        public virtual double LuminousEfficacy { get; set; } = 0.0;
+        Undefined,
+        Absolute,
+        ModulatingPercentage,
+        OnOff
     }
 }
