+++ Environment_oM/SpaceCriteria/Profile.cs
@@ -27,12 +27,28 @@ using System.Text;
 using System.Threading.Tasks;
 
 using BH.oM.Base;
+using System.ComponentModel;
+using BH.oM.Environment.Fragments;
 
-namespace BH.oM.Environment.Gains
+namespace BH.oM.Environment.SpaceCriteria
 {
-    public interface IGain : IBHoMObject
+    [Description("A profile object represents a gain, thermostat or humdistat as conditions vary over a 24-hour period")]
+    public class Profile : BHoMObject, IEnvironmentObject
     {
+        /***************************************************/
+        /**** Properties                                ****/
+        /***************************************************/
 
+        [Description("Profile Type depicts the type of profile (OnOff (0 or 1), ModulatingPercentage (0.0-1.0), or Absolute (-10000 to +10000))")]
+        public virtual ProfileType ProfileType { get; set; } = ProfileType.Undefined;
+
+        [Description("Profile Day depicts the day described in the profile, whether it's a day of the week or a holiday")]
+        public virtual List<ProfileDay> ProfileDay { get; set; } = new List<ProfileDay>();
+
+        [Description("Hourly Values denotes the hour-by-hour values for a 24-hour period. These values may be represented in temperature (degrees C) (thermostat), fraction (0.9) (lighting gain), or flow (m3/s) (ventilation)")]
+        public virtual List<double> HourlyValues { get; set; } = new List<double>();
+
+        /***************************************************/
     }
 }
